Updates and bugfixes.
Removed redundant or useless code from profile.js and html. Updated titles and meta tags for all html pages. Fixed issue of more button not working half the time, and issues with restaurant loading the skeleton and some functions out of order. Updated skeleton to force event listeners to load in after the rest of the html is loaded.